Organize Document classes
Add FullTextDocument  class to separate documents which can have full text (eg. Mets) and can not (3D model)
