+++ Classes/Plugin/TableOfContents.php
@@ -41,7 +41,7 @@ class TableOfContents extends \Kitodo\Dlf\Common\AbstractPlugin
      *
      * @access protected
      *
-     * @param array $entry: The entry's array from \Kitodo\Dlf\Common\Document->getLogicalStructure
+     * @param array $entry: The entry's array from \Kitodo\Dlf\Common\Document\Document->getLogicalStructure
      * @param bool $recursive: Whether to include the child entries
      *
      * @return array HMENU array for menu entry
